Rename file to match the exported identifier
