+++ packages/mui-base/src/FormControlUnstyled/useFormControlUnstyledContext.ts
@@ -1,5 +1,5 @@
 import * as React from 'react';
-import FormControlUnstyledContext from './FormControlContext';
+import FormControlUnstyledContext from './FormControlUnstyledContext';
 
 export default function useFormControlUnstyledContext() {
   return React.useContext(FormControlUnstyledContext);
